requirements.txt
termcolor
groq
python-dotenv
huggingface-hub
praw
bs4
requests
gradio
pandas
plotly
folium
google-generativeai
mistralai